don't forget to connect the framebuffers to the connector
Each new framebuffer has to be connected to the connector. Otherwise,
the panel may not be enabled.
With libdrm < 2.4.99, there is an extra open/close of the DRM file which
hides this missing part. When the file is closed, there a kind of reset
mecanism which setup the connector in the proper way.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>