reset the state of a plane when requesting a new one
The lifetime of a plane is not the same in libplanes and in the kernel.
In the kernel, planes are initialized when the lcd driver is probed,
there is not reset of their state that is done except when suspending
and resuming the lcd driver.
With the abstraction provided by libplanes, a plane is seen as an object
that is requested, used and released later. We could assume that the
state of the plane is resetted at some point, but it's not the case. The
properties of the hardware plane could have been modified, but the kernel
has no clue that a software plane is released and so the property changes
should be reverted.
So, when requesting a software plane, set the alpha and rotation
properties to a default value in order to start with a known and obvious
state.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>