add CMake support
Add CMake support for the build system.

Take this opportunity to rework libplanes hard dependencies. The
libplanes engine has many dependencies that are not necessary for
the library. Moreover, this engine is mostly a demonstrator or a
debug tool, so we should be able to get rid of it if we just want
to use libplanes as a library only. The drawing and rendering
supports in libplanes are mostly for the engine, so enable this part
of the library only if we want the engine.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Reviewed-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>