switch to the DRM atomic API
Stop using the legacy DRM API and switch to the atomic one. This move
involves gathering the state changes before committing them. It means
a new public function needs to be introduced to commit the changes. It
is the kms_device_flush() function. If not called, the changes are not
committed and so nothing happens.

Applications using libplanes must add the call to kms_device_flush()
when they want to update the display.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>